use postgres-compatible alias for double "float8"

+++ db/migrate/20110519150816_update_locations_floats_to_doubles.rb
@@ -1,8 +1,8 @@
 class UpdateLocationsFloatsToDoubles < ActiveRecord::Migration
   def self.up
     rename_column :locations, :long, :lon
-    change_column :locations, :lat, :double
-    change_column :locations, :lon, :double
+    change_column :locations, :lat, :float8
+    change_column :locations, :lon, :float8
   end
 
   def self.down
